TOOLS-166 rework command structure
