+++ cmd/main.go
@@ -23,8 +23,8 @@ import (
 )
 
 func main() {
-	resp := commands.Execute(os.Args[1:])
-	if resp.Err != nil {
-		pterm.Error.Print(resp.Err)
+	err := commands.Execute(os.Args[1:])
+	if err != nil {
+		pterm.Error.Print(err)
 	}
 }
