--- internal/commands/comandeer.go
@@ -1,7 +1,0 @@
-package commands
-
-import "github.com/hazelops/ize/pkg/logger"
-
-type commandeer struct {
-	logger logger.StandartLogger
-}
