+++ internal/config/helpers_test.go
@@ -1,4 +1,4 @@
-package commands
+package config
 
 import "testing"
